split to headers and sources
