+++ src/headers/ability.h
@@ -12,21 +12,15 @@ struct Ability
 {
     explicit Ability(
         std::string name = "unnamed",
-        const uint32_t hp_effect = 1u,
+        uint32_t hp_effect = 1u,
         Buff* buff = nullptr,
-        const uint32_t cost = 0u,
-        const uint32_t cooldown = 1u,
-        const AbilityTarget target = AbilityTarget::self,
-        const AbilityScaler scaler = AbilityScaler::none
-    ):
-    name(std::move(name)),
-    hp_effect(hp_effect),
-    given_buff(buff),
-    cost(cost),
-    cooldown(cooldown),
-    target(target),
-    scaler(scaler)
-    {}
+        uint32_t cost = 0u,
+        uint32_t cooldown = 1u,
+        AbilityTarget target = AbilityTarget::self,
+        AbilityScaler scaler = AbilityScaler::none
+    );
+
+    ~Ability();
 
     std::string name;
     Buff* given_buff;
