+++ src/headers/buff.h
@@ -8,36 +8,27 @@
 
 struct Buff
 {
+    Buff();
+
     explicit Buff (
         const char* name = "undefined",
-        const uint16_t duration = 2u,
-        const CoreStats stats = CoreStats(),
-        const bool is_debuff = false
-    ): name(name), duration(duration), stats(stats), is_debuff(is_debuff) {}
+        uint16_t duration = 2u,
+        CoreStats stats = CoreStats(),
+        bool is_debuff = false
+    );
 
     explicit Buff(
         const char* name,
-        const uint16_t duration = 2u,
-        const stattype strength = 0,
-        const stattype intellect = 0,
-        const stattype agility = 0,
-        const stattype armor = 0,
-        const stattype resistance = 0,
-        const bool is_debuff = false
-    ):
-    name(name),
-    duration(duration),
-    stats({strength, intellect, agility, armor, resistance}),
-    is_debuff(is_debuff) {}
+        uint16_t duration = 2u,
+        stattype strength = 0,
+        stattype intellect = 0,
+        stattype agility = 0,
+        stattype armor = 0,
+        stattype resistance = 0,
+        bool is_debuff = false
+    );
 
-    Buff& operator=(const char* name)
-    {
-        this->name = std::string(name);
-        this->duration = 2u;
-        this->stats = CoreStats(0);
-        this->is_debuff = false;
-        return *this;
-    }
+    Buff& operator=(const char* name);
 
     std::string name;
     uint16_t duration;
