+++ src/headers/level_system.h
@@ -6,35 +6,17 @@
 class LevelSystem
 {
 public:
-    LevelSystem()
-    {
-        current_level = 1u;
-        current_exp = 0u;
-        exp_to_next_level = kLevel2At;
-    }
+    LevelSystem();
 
     virtual ~LevelSystem() = default;
 
-    void gain_exp(const exptype gained_exp)
-    {
-        current_exp += gained_exp;
-        check_if_leveled();
-    }
+    void gain_exp(exptype gained_exp);
 
-    [[nodiscard]] leveltype get_level() const
-    {
-        return current_level;
-    }
+    [[nodiscard]] leveltype get_level() const;
 
-    [[nodiscard]] exptype get_current_exp() const
-    {
-        return current_exp;
-    }
+    [[nodiscard]] exptype get_current_exp() const;
 
-    [[nodiscard]] exptype get_exp_to_next_level() const
-    {
-        return exp_to_next_level;
-    }
+    [[nodiscard]] exptype get_exp_to_next_level() const;
 
     virtual void level_up() = 0;
