+++ src/tests.cpp
@@ -3,7 +3,10 @@
 #include <vector>
 
 #include "headers/item_manager.h"
-#include "headers/player_classes.h"
+#include "headers/cleric.h"
+#include "headers/rogue.h"
+#include "headers/warrior.h"
+#include "headers/wizard.h"
 
 void separator(const std::string& func_name, const int& line)
 {
@@ -274,7 +277,8 @@ void test_inventory()
         const auto inventory = wizard.get_backpack_list();
         std::cout << "Inventory: ";
         for (const auto it: inventory)
-            std::cout << *it << ", ";
+            // std::cout << *it << ", ";
+            std::cout << it->get_data() << ", ";
         std::cout << "\n\n";
     }
 
@@ -301,7 +305,8 @@ void test_inventory()
         const auto inventory = wizard.get_backpack_list();
         std::cout << "\nInventory after healing: ";
         for (const auto it: inventory)
-            std::cout << *it << ", ";
+            // std::cout << *it << ", ";
+            std::cout << it->get_data() << ", ";
         std::cout << '\n';
     }
 }
